+++ docs/examples/searchguide/avatar.tsx
