+++ docs/examples/searchguide/expandable.tsx
