+++ docs/examples/searchguide/selectedState.tsx
