+++ docs/examples/searchguidelink/avatar.tsx
