+++ docs/pages/web/searchguide.tsx
@@ -8,13 +8,13 @@ import Page from '../../docs-components/Page';
 import PageHeader from '../../docs-components/PageHeader';
 import QualityChecklist from '../../docs-components/QualityChecklist';
 import SandpackExample from '../../docs-components/SandpackExample';
-import avatarExample from '../../examples/searchguide/avatarExample';
+import avatarExample from '../../examples/searchguide/avatar';
 import colors from '../../examples/searchguide/colors';
-import expandableExample from '../../examples/searchguide/expandableExample';
+import expandableExample from '../../examples/searchguide/expandable';
 import hideLowQualityDo from '../../examples/searchguide/hideLowQualityDo';
 import main from '../../examples/searchguide/main';
 import noScrollDo from '../../examples/searchguide/noScrollDo';
-import selectedStateExample from '../../examples/searchguide/selectedStateExample';
+import selectedStateExample from '../../examples/searchguide/selectedState';
 import serialDo from '../../examples/searchguide/serialDo';
 import showLowQualityDont from '../../examples/searchguide/showLowQualityDont';
 import singleDont from '../../examples/searchguide/singleDont';
